Moving MorphModel logic to trait.
